python-fastapi-jwt/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.92.0
h11==0.14.0
idna==3.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
